requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2024.6.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
django==5.0.6; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
django-environ==0.11.2; python_version >= '3.6' and python_version < '4'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-yasg==1.21.7; python_version >= '3.6'
idna==3.7; python_version >= '3.5'
inflection==0.5.1; python_version >= '3.5'
packaging==24.1; python_version >= '3.8'
pillow==10.3.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
requests==2.32.3; python_version >= '3.0'
sqlparse==0.5.0; python_version >= '3.8'
stripe==9.11.0; python_version >= '3.6'
typing-extensions==4.12.2; python_version >= '3.7'
tzdata==2024.1; sys_platform == 'win32'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.2; python_version >= '3.8'
whitenoise==6.6.0; python_version >= '3.8'
